legal: change license to GPLv3 in script headers
Kept wondering wether I should keep the MIT (because I highly value
freedom of act) or embrace the GPL (because I don't want this work to
become close-source).

I do understand that this commit (and the next one that is actually
changing the LICENSE) is a defeat for freedom. I guess freedom has been
defeated long time ago, when people and companies figured that was
"free" (as in gratis) was also "free" (as in disposable).

This is not how I think free (as in "libre") works but hey, that surely
is how Intel and other corporations see it (ec: Intel Management Engine
is entirely based on Minix, is close-source, and *maybe* used as a
backdoor by anybody).

It boils down to the Paradox of Tolerance, and I surely won't tolerate
shit going their way. If you want to take open source stuff, be my
guest ; but you have to play by the rules.